Changed how the game's hooks are installed to when the game is already running.
This is done because the game's latest update encrypted its executable with Steam's DRM, which makes it impossible to install hooks before Steam decrypts the game's code.
The safest way that I found so far is to hook on the game's update code when the game calls the dinput's DirectInput8Create function. The Update function hook will then take care of initializing the rest of the tool when it is first called.